requirements.txt
django>=3.2
djangorestframework>=3.12.4
requests>=2.25.1